Rename agent_core -> agent
