+++ Cargo.toml
@@ -3,7 +3,7 @@ resolver = "2"
 members = [
     "crates/activity_indicator",
     "crates/agent_ui",
-    "crates/agent_core",
+    "crates/agent",
     "crates/agent_settings",
     "crates/anthropic",
     "crates/askpass",
@@ -214,8 +214,8 @@ edition = "2024"
 #
 
 activity_indicator = { path = "crates/activity_indicator" }
+agent = { path = "crates/agent" }
 agent_ui = { path = "crates/agent_ui" }
-agent_core = { path = "crates/agent_core" }
 agent_settings = { path = "crates/agent_settings" }
 ai = { path = "crates/ai" }
 anthropic = { path = "crates/anthropic" }
