+++ crates/agent/src/history_store.rs
