+++ crates/agent/src/prompts/summarize_thread_detailed_prompt.txt
