+++ crates/agent/src/prompts/summarize_thread_prompt.txt
