+++ crates/agent/src/thread_store.rs
