+++ crates/agent_ui/src/agent_configuration/manage_profiles_modal.rs
@@ -16,7 +16,7 @@ use workspace::{ModalView, Workspace};
 use crate::agent_configuration::manage_profiles_modal::profile_modal_header::ProfileModalHeader;
 use crate::agent_configuration::tool_picker::{ToolPicker, ToolPickerDelegate};
 use crate::{AgentPanel, ManageProfiles};
-use agent_core::agent_profile::AgentProfile;
+use agent::agent_profile::AgentProfile;
 
 use super::tool_picker::ToolPickerMode;
