+++ crates/agent_ui/src/agent_ui.rs
@@ -21,7 +21,7 @@ mod ui;
 
 use std::sync::Arc;
 
-use agent_core::{Thread, ThreadId};
+use agent::{Thread, ThreadId};
 use agent_settings::{AgentProfileId, AgentSettings, LanguageModelSelection};
 use assistant_slash_command::SlashCommandRegistry;
 use client::Client;
@@ -148,7 +148,7 @@ pub fn init(
         init_language_model_settings(cx);
     }
     assistant_slash_command::init(cx);
-    agent_core::init(cx);
+    agent::init(cx);
     agent_panel::init(cx);
     context_server_configuration::init(language_registry.clone(), fs.clone(), cx);
