+++ crates/agent_ui/src/context_picker/rules_context_picker.rs
@@ -8,8 +8,8 @@ use ui::{ListItem, prelude::*};
 use util::ResultExt as _;
 
 use crate::context_picker::ContextPicker;
-use agent_core::context::RULES_ICON;
-use agent_core::context_store::{self, ContextStore};
+use agent::context::RULES_ICON;
+use agent::context_store::{self, ContextStore};
 
 pub struct RulesContextPicker {
     picker: Entity<Picker<RulesContextPickerDelegate>>,
