+++ crates/eval/src/instance.rs
@@ -1,4 +1,4 @@
-use agent_core::{Message, MessageSegment, SerializedThread, ThreadStore};
+use agent::{Message, MessageSegment, SerializedThread, ThreadStore};
 use anyhow::{Context as _, Result, anyhow, bail};
 use assistant_tool::ToolWorkingSet;
 use client::proto::LspWorkProgress;
