+++ crates/zed/Cargo.toml
@@ -20,8 +20,8 @@ path = "src/main.rs"
 
 [dependencies]
 activity_indicator.workspace = true
+agent.workspace = true
 agent_ui.workspace = true
-agent_core.workspace = true
 agent_settings.workspace = true
 anyhow.workspace = true
 askpass.workspace = true
